+++ file-transfer/Solicitud.cpp
@@ -21,7 +21,7 @@
 
 /* TODO(Garz4): Consider making Solicitud constexpr, moving this to header. */
 const Mensaje& Solicitud::envia_y_recibe(
-      const std::string& ip, int puerto, const Mensaje& solicitud) {
+      const std::string& ip, unsigned short puerto, const Mensaje& solicitud) {
   enviar_ = solicitud;
   enviar_.tipo = '0';
   enviar_.id = peticion_;
